Delete old migration code
