+++ runtime/common/src/lib.rs
@@ -27,7 +27,6 @@ pub mod impls;
 pub mod paras_registrar;
 pub mod paras_sudo_wrapper;
 pub mod purchase;
-pub mod session;
 pub mod slot_range;
 pub mod slots;
 pub mod traits;
