+++ runtime/parachains/src/configuration/migration.rs
@@ -16,8 +16,5 @@
 
 //! A module that is responsible for migration of storage.
 
-use primitives::ExecutorParams;
-
-pub mod v5;
 pub mod v6;
 pub mod v7;
